small change -- removed some redundant code

+++ Scripts/PCG/BlockPlacer.cs
@@ -159,7 +159,7 @@ public partial class BlockPlacer : Area2D
 			// select the minimum amount of rooms based on the current difficulty settin
 			int min_parts = 3;
 			int max_parts = (int)Math.Floor(difficulty/3) + 3;
-			num_parts_in_room = (int)random.Next(min_parts, max_parts+1); 
+			num_parts_in_room = random.Next(min_parts, max_parts+1); 
 		} 
 		else { 
 			num_parts_in_room = num_parts;
